Add react components to the passage result context message
