+++ src/utils/getPassageResultsContext.tsx
@@ -1,0 +1,39 @@
+import { ConceptLink } from "@/components/molecules/conceptLink/ConceptLink";
+import { MAX_PASSAGES } from "@/constants/paging";
+import { TConcept } from "@/types";
+
+interface IArgs {
+  isExactSearch?: boolean;
+  passageMatches: number;
+  queryTerm: string | string[];
+  selectedTopics: TConcept[];
+}
+
+export const getPassageResultsContext = ({ isExactSearch, passageMatches, queryTerm, selectedTopics }: IArgs) => {
+  const queryString = queryTerm instanceof Array ? queryTerm[0] || "" : queryTerm;
+
+  const passages =
+    passageMatches >= MAX_PASSAGES ? (
+      <>
+        Top <span className="font-bold">{MAX_PASSAGES}</span>
+      </>
+    ) : (
+      <span className="font-bold">{passageMatches}</span>
+    );
+  const matchesPlural = passageMatches === 1 ? "match" : "matches";
+
+  const phrase = queryString ? `${isExactSearch ? "" : "phrases related to "}"${queryString}"` : null;
+
+  return (
+    <div>
+      {passages} {matchesPlural} for {phrase}{" "}
+      {selectedTopics.map((concept, i, allConcepts) => (
+        <>
+          {i === 0 && "AND "}
+          <ConceptLink key={concept.wikibase_id} concept={concept} />
+          {i + 1 < allConcepts.length && " AND "}
+        </>
+      ))}
+    </div>
+  );
+};
